add origin

+++ src/main/java/edu/muc/jxd/cluster/Cluster.java
@@ -15,7 +15,6 @@ public class Cluster {
 	private List<ImageItemVector<Number>> itemList;
 
 	private List<Result> resultList;
-	
 
 	/**
 	 * 用来计算小兵的肉
@@ -32,7 +31,7 @@ public class Cluster {
 	 * 
 	 * @param items
 	 */
-	public Cluster(List<ImageItemVector<Number>> items, DistenceInter distance, int start, int end, int step) {
+	public Cluster(List<ImageItemVector<Number>> items, DistenceInter distance, double start, double end, double step) {
 		this.itemList = items;
 		this.resultList = new ArrayList<Result>();
 		/*
@@ -61,7 +60,8 @@ public class Cluster {
 		Collections.sort(this.resultList);
 		StringBuilder builder = new StringBuilder();
 		for (Result result : this.resultList) {
-			builder.append(result.getId() + " : "+"p: "+result.getP()+" D: "+result.getDelta() +" "+ result.getMutiplyResult() + "\n");
+			builder.append(result.getId() + " : " + "p: " + result.getP() + " D: " + result.getDelta() + " "
+					+ result.getMutiplyResult() + "\n");
 		}
 		System.out.println(builder.toString());
 	}
@@ -89,6 +89,5 @@ public class Cluster {
 	public void setDelta(Delta delta) {
 		this.delta = delta;
 	}
-	
 
 }
